requirements.txt
PyQt5>=5.15.0
opencv-python>=4.5.0
face-recognition>=1.3.0
mediapipe>=0.8.0
matplotlib>=3.4.0
openai>=0.27.0
python-dotenv>=0.19.0
numpy>=1.19.0
pymongo>=4.3.0
pyaudio>=0.2.13
wave>=0.0.2
certifi
SpeechRecognition 